Updated database model and created new controllers

+++ Controllers/CommentsController.cs
@@ -30,7 +30,7 @@ namespace NoCom_API.Controllers
 
         // GET: api/Comments/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Comment>> GetComment(int id)
+        public async Task<ActionResult<Comment>> GetComment(long id)
         {
             var comment = await _context.Comments.FindAsync(id);
 
@@ -45,7 +45,7 @@ namespace NoCom_API.Controllers
         // PUT: api/Comments/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutComment(int id, Comment comment)
+        public async Task<IActionResult> PutComment(long id, Comment comment)
         {
             if (id != comment.Id)
             {
@@ -100,7 +100,7 @@ namespace NoCom_API.Controllers
 
         // DELETE: api/Comments/5
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteComment(int id)
+        public async Task<IActionResult> DeleteComment(long id)
         {
             var comment = await _context.Comments.FindAsync(id);
             if (comment == null)
@@ -114,7 +114,7 @@ namespace NoCom_API.Controllers
             return NoContent();
         }
 
-        private bool CommentExists(int id)
+        private bool CommentExists(long id)
         {
             return _context.Comments.Any(e => e.Id == id);
         }
