+++ Models/LikedComment.cs
@@ -8,9 +8,9 @@ namespace NoCom_API.Models
     /// </summary>
     public partial class LikedComment
     {
-        public int Id { get; set; }
-        public int UserId { get; set; }
-        public int CommentId { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; }
+        public long CommentId { get; set; }
 
         public virtual Comment Comment { get; set; } = null!;
         public virtual User User { get; set; } = null!;
